refactor(web): dedupe OrderSummary type into shared/types
The OrderSummary interface was declared identically in both
shipments.types.ts and invoicing.types.ts, mirroring the same backend
OrderSummaryProjectionDto with no natural owning feature. Move it to
shared/types/order-summary.types.ts (matching the structured-error.types.ts
precedent) so both features import the same definition instead of two
independently-drifting copies.

Closes #2020
